Fix incorrect consumption of chars for token
Switch to use a..<b proc in order to exclude upper bound.
When using a..b, it is inclusive of b. This resulted in
recognizing (( as a LEFT_PAREN or {{ as a LEFT_BRACKET.

In the Java example from the book, the substring() method was
used which begins at the specified beginIndex and extends to the
character at index endIndex - 1. In other words, it is a half-open
interval, i.e. exclusive of the endIndex.

Modify our code to be in line with this and fix the incorrect
tokenizing.